pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["VictorLoureiro <v.loureiro11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.10"
jupyter = "^1.0.0"
requests = "^2.28.2"
pandas = "^1.5.3"
pyarrow = "^11.0.0"
tqdm = "^4.64.1"
plotly = "^5.13.1"
scikit-learn = "^1.2.1"
lightgbm = "^3.3.5"
xgboost = "^1.7.4"
optuna = "^3.1.0"
python-dotenv = "^1.0.0"
hopsworks = "^3.0.5"
streamlit = "^1.19.0"
geopandas = "^0.12.2"
pydeck = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
